ViewModelのテスト修正
